fix: make csp middleware url specific (#35318)
